requirements.txt
livekit-agents[deepgram,silero]>=0.8.0
python-dotenv>=1.0.0